requirements.txt
tensorflow
keras
sklearn
pandas
numpy
seaborn
visualkeras
streamlit
python-dotenv